+++ forJIS7/src/main/java/JavaHomework/Lectures_17.Stream/secondPart/Repository/UserRepository.java
@@ -8,4 +8,6 @@ import java.util.List;
 public class UserRepository {
 
     public static List<User> userList = new ArrayList<>();
+
+
 }
